requirements.txt
flask
line-bot-sdk
openai
langchain-community
langchain-google-genai
chromadb
pysqlite3-binary
gunicorn